refactor: ChargerList 분리, StationInformation 분리
[#435]
