+++ frontend/src/components/ui/StationDetailsWindow/chargers/ChargerCard.tsx
