+++ frontend/src/components/ui/StationDetailsWindow/chargers/ChargerCardSkeleton.tsx
