+++ frontend/src/components/ui/StationDetailsWindow/chargers/ChargerList.tsx
@@ -1,0 +1,21 @@
+import FlexBox from '@common/FlexBox';
+
+import ChargerCard from '@ui/StationDetailsWindow/chargers/ChargerCard';
+
+import type { Charger } from '@type';
+
+export interface ChargerListProps {
+  chargers: Charger[];
+}
+
+const ChargerList = ({ chargers }: ChargerListProps) => {
+  return (
+    <FlexBox>
+      {chargers.map((charger, index) => (
+        <ChargerCard key={index} charger={charger} />
+      ))}
+    </FlexBox>
+  );
+};
+
+export default ChargerList;
