+++ frontend/src/components/ui/StationDetailsWindow/reports/StationReportPreConfirmation.tsx
@@ -7,8 +7,8 @@ import ButtonNext from '@common/ButtonNext';
 import FlexBox from '@common/FlexBox';
 import Text from '@common/Text';
 
-import StationInformation from '@ui/StationDetailsWindow/StationInformation';
 import StationReportConfirmation from '@ui/StationDetailsWindow/reports/StationReportConfirmation';
+import StationInformation from '@ui/StationDetailsWindow/station/StationInformation';
 
 import type { StationDetails } from '@type';
