+++ frontend/src/components/ui/StationDetailsWindow/station/StationInformation.tsx
