+++ frontend/src/components/ui/StationDetailsWindow/station/StationInformationSkeleton.tsx
